Migrate `MapFile` class into separate translation unit
No reason for this to be in `Blam.hpp`
